ルーターをモジュールに切り出す（vanilla JSにて）
