+++ index_dev.html
@@ -12,7 +12,7 @@
   <!-- JavaScriptインポート -->
   <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.24.0/moment.min.js" defer></script><!-- Moment.js -->
   <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.24.0/locale/ja.js" defer></script><!-- Moment.js 日本語ロケール -->
-  <script src="./src/main.js" defer></script><!-- ページ生成スクリプト -->
+  <script type="module" src="./src/index.js"></script><!-- ページ生成スクリプト -->
   <!-- アイコン画像 -->
   <link rel="icon" href="./assets/icons/favicon.svg" type="image/svg+xml" />
   <link rel="apple-touch-icon" type="image/png" sizes="180x180" href="./assets/icons/apple-touch-icon-180x180.png" />
