+++ src/__tests__/router.test.ts
@@ -1,0 +1,6 @@
+import { getUrlQueries } from '../index';
+
+test('複数のクエリ文字列', () => {
+  expect(getUrlQueries('foo=1&bar=2'))
+    .toBe('');
+});
