+++ src/index.js
@@ -1,32 +1,12 @@
 'use strict'
-{  
+import { getUrlQueries } from './router.js';
+
   // 表示調整用
   const jsonPath = './data.json';
 
-/* ---------------------------------
-  URLからクエリ文字列を取得 */
-  const getUrlQueries = () => {
-    const queries = {};
-    const queryStr = window.location.search.slice(1);  // 'foo=1&bar=2'、文頭の'?'を除外
-  
-    // クエリがない場合は、
-    if (!queryStr) {
-      return queries;  // 空のオブジェクトを返す。
-    }
-    
-    // 複数のクエリを'&'で切って配列へと分解
-    const queryArr = queryStr.split('&')  // ['foo=1', 'bar=2']
-    queryArr.forEach((eachQueryStr) => {
-      const keyAndValue = eachQueryStr.split('=');  // ['foo', '1']// '='でさらに分割してそれぞれ配列（key,value）へと格納
-      // 配列からオブジェクトを生成、このとき値を日本語にデコードしておく
-      queries[keyAndValue[0]] = decodeURIComponent(keyAndValue[1]);  // {foo: 1}
-    });
-  
-    return queries;
-  };
-  
   // 実行
-  const status = getUrlQueries();
+  const queryStr = window.location.search.slice(1);  // 'foo=1&bar=2'、文頭の'?'を除外
+  const status = getUrlQueries(queryStr);
   
 /* JSONデータ取得開始 --------------------  */
 fetch(jsonPath)
@@ -313,5 +293,3 @@ fetch(jsonPath)
   }
 
 })  // fetch.then((data) => {...
-
-} // {...
